python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0
opencv-python==4.8.0.76
mediapipe==0.10.7 